+++ src/nsidc/iceflow/data/read.py
@@ -6,7 +6,11 @@ import pandas as pd
 
 from nsidc.iceflow.data.atm1b import atm1b_data
 from nsidc.iceflow.data.glah06 import glah06_data
-from nsidc.iceflow.data.ilvis2 import ilvis2_data
+from nsidc.iceflow.data.ilvis2 import (
+    ILVIS2_COORDINATE_SETS,
+    ILVIS2_DEFAULT_COORDINATE_SET,
+    ilvis2_data,
+)
 from nsidc.iceflow.data.models import (
     ATM1BDataFrame,
     GLAH06DataFrame,
@@ -17,7 +21,9 @@ from nsidc.iceflow.data.models import (
 
 def read_iceflow_datafile(
     filepath: Path,
+    ilvis2_coordinate_set: ILVIS2_COORDINATE_SETS = ILVIS2_DEFAULT_COORDINATE_SET,
 ) -> IceflowDataFrame | ATM1BDataFrame | ILVIS2DataFrame | GLAH06DataFrame:
+    """Read the given iceflow datafile."""
     # iceflow data are expected to exist in a directory named like
     # `{short_name}_{version}`
     dataset_subdir = filepath.parent.name
@@ -26,7 +32,7 @@ def read_iceflow_datafile(
     if short_name in ["ILATM1B", "BLATM1B"]:
         return atm1b_data(filepath)
     elif short_name == "ILVIS2":
-        return ilvis2_data(filepath)
+        return ilvis2_data(filepath, coordinate_set=ilvis2_coordinate_set)
     elif short_name == "GLAH06":
         return glah06_data(filepath)
     else:
@@ -34,10 +40,16 @@ def read_iceflow_datafile(
         raise RuntimeError(err_msg)
 
 
-def read_iceflow_datafiles(filepaths: list[Path]) -> IceflowDataFrame:
+def read_iceflow_datafiles(
+    filepaths: list[Path],
+    ilvis2_coordinate_set: ILVIS2_COORDINATE_SETS = ILVIS2_DEFAULT_COORDINATE_SET,
+) -> IceflowDataFrame:
     all_dfs = []
     for filepath in filepaths:
-        df = read_iceflow_datafile(filepath)
+        df = read_iceflow_datafile(
+            filepath,
+            ilvis2_coordinate_set=ilvis2_coordinate_set,
+        )
         all_dfs.append(df)
 
     complete_df = IceflowDataFrame(pd.concat(all_dfs))